backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
jose==1.0.0
passlib==1.7.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
